Move `createThemeValuesProxy` function to seperate file
