+++ packages/theme/src/__tests__/createThemeValuesProxy.test.ts
@@ -1,4 +1,4 @@
-import { createThemeValuesProxy } from "../theming";
+import createThemeValuesProxy from "../createThemeValuesProxy";
 
 const breakpoints = {
   small: 100,
